requirements.txt
flask>=1.1.0
flask-Bootstrap4>=4.0.0
flask-WTF>=0.14.2
